requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
pytest
dateparser
